+++ mozapkpublisher/test/test_update_apk_description.py
@@ -8,6 +8,8 @@ from unittest.mock import create_autospec
 from tempfile import NamedTemporaryFile
 
 from mozapkpublisher.common import googleplay, store_l10n
+from mozapkpublisher.common.exceptions import WrongArgumentGiven
+from mozapkpublisher.common.googleplay import MockGooglePlayConnection
 from mozapkpublisher.update_apk_description import main, update_apk_description
 
 
@@ -27,7 +29,7 @@ def test_update_apk_description_force_locale(monkeypatch):
     })
     monkeypatch.setattr(store_l10n, '_translate_moz_locate_into_google_play_one', lambda locale: 'google_play_locale')
 
-    update_apk_description('org.mozilla.firefox_beta', 'en-US', False, 'foo@developer.gserviceaccount.com', credentials, False)
+    update_apk_description(MockGooglePlayConnection(), 'org.mozilla.firefox_beta', 'en-US', False)
 
     google_play_mock.update_listings.assert_called_once_with(
         'google_play_locale',
@@ -41,13 +43,14 @@ def test_update_apk_description_force_locale(monkeypatch):
 
 def test_main(monkeypatch):
     incomplete_args = [
+        'script',
         '--package-name', 'org.mozilla.firefox_beta',
         '--service-account', 'foo@developer.gserviceaccount.com',
     ]
 
     monkeypatch.setattr(sys, 'argv', incomplete_args)
 
-    with pytest.raises(SystemExit):
+    with pytest.raises(WrongArgumentGiven):
         main()
 
     complete_args = [
@@ -57,5 +60,6 @@ def test_main(monkeypatch):
         '--credentials', os.path.join(os.path.dirname(__file__), 'data', 'blob')
     ]
     monkeypatch.setattr(sys, 'argv', complete_args)
-    monkeypatch.setattr(mozapkpublisher.update_apk_description, 'update_apk_description', lambda _, __, ___, ____, _____, ______: None)
+    monkeypatch.setattr(mozapkpublisher.update_apk_description, 'update_apk_description', lambda _, __, ___, ____: None)
+    monkeypatch.setattr(mozapkpublisher.update_apk_description, 'connection_for_options', lambda _, __, ___: None)
     main()